# Remove Roslyn semantic tokens edit handling + switch to Roslyn range handling