--- a/src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer.Common/LanguageServerConstants.cs
+++ b/src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer.Common/LanguageServerConstants.cs
@@ -42,7 +42,7 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Common
 
         public const string RazorResolveCodeActionsEndpoint = "razor/resolveCodeActions";
 
-        public const string RazorProvideSemanticTokensEndpoint = "razor/provideSemanticTokens";
+        public const string RazorProvideSemanticTokensRangeEndpoint = "razor/provideSemanticTokensRange";
 
         public const string RazorServerReadyEndpoint = "razor/serverReady";