--- a/src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Semantic/Models/ProvideSemanticTokensParams.cs
+++ b/src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Semantic/Models/ProvideSemanticTokensParams.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT license. See License.txt in the project root for license information.
 
+#nullable enable
+
 using OmniSharp.Extensions.LanguageServer.Protocol.Models;
 
 namespace Microsoft.AspNetCore.Razor.LanguageServer.Semantic.Models
@@ -8,5 +10,7 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Semantic.Models
     public record ProvideSemanticTokensParams : SemanticTokensParams
     {
         public long RequiredHostDocumentVersion { get; set; }
+
+        public Range? Range { get; set; }
     }
 }