… src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Semantic/Models/ProvideSemanticTokensDeltaParams.cs
… src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Semantic/Models/ProvideSemanticTokensEditsResponse.cs
… src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Semantic/DefaultRazorSemanticTokenInfoServiceTest.cs
… src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Semantic/TestFiles/DefaultRazorSemanticTokenInfoServiceTest/ApplyEditsToPreviousCSharpDoc_EnsureAccurateDeletionEdits.semanticedit.txt
… src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Semantic/TestFiles/DefaultRazorSemanticTokenInfoServiceTest/ApplyEditsToPreviousCSharpDoc_EnsureAccurateDeletionEdits.semanticexpected.txt
… src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Semantic/TestFiles/DefaultRazorSemanticTokenInfoServiceTest/ApplyEditsToPreviousCSharpDoc_EnsureAccurateDeletionEdits.semanticoriginal.txt
… src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Semantic/TestFiles/DefaultRazorSemanticTokenInfoServiceTest/ApplyEditsToPreviousCSharpDoc_EnsureAccurateInsertionDeletionEdits.semanticedit.txt
… src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Semantic/TestFiles/DefaultRazorSemanticTokenInfoServiceTest/ApplyEditsToPreviousCSharpDoc_EnsureAccurateInsertionDeletionEdits.semanticexpected.txt
… src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Semantic/TestFiles/DefaultRazorSemanticTokenInfoServiceTest/ApplyEditsToPreviousCSharpDoc_EnsureAccurateInsertionDeletionEdits.semanticoriginal.txt
… src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Semantic/TestFiles/DefaultRazorSemanticTokenInfoServiceTest/ApplyEditsToPreviousCSharpDoc_EnsureAccurateInsertionEdits.semanticedit.txt
… src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Semantic/TestFiles/DefaultRazorSemanticTokenInfoServiceTest/ApplyEditsToPreviousCSharpDoc_EnsureAccurateInsertionEdits.semanticexpected.txt
… src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Semantic/TestFiles/DefaultRazorSemanticTokenInfoServiceTest/ApplyEditsToPreviousCSharpDoc_EnsureAccurateInsertionEdits.semanticoriginal.txt
… src/Razor/test/Microsoft.VisualStudio.LanguageServerClient.Razor.Test/DefaultRazorLanguageServerCustomMessageTargetTest.cs