### Summary of the changes
 - Roslyn is attempting to deprecate its edits handling experience, which means Razor needs to switch off calling into the Roslyn edits handler. This PR changes Razor's code so now it only calls into Roslyn's range handler and no longer calls into the full document and edits handler.
 - The PR makes an optimization so we only ask Roslyn for a minimal range in the document, i.e. only colorizes the range spanning from the first C# span to the last C# span.
 - Essentially perf-wise this PR has minimal benefit to Razor and mostly benefits Roslyn, however I'm planning on investigating other range optimizations in the future (e.g. range partitioning on the Roslyn side) that should improve perf further. Current range perf is slightly lower than edits handling but it doesn't seem noticeable in actual scenarios. All of these numbers are from debug mode so they aren't quite accurate overall but should be accurate relative to one another:

<html xmlns:o="urn:schemas-microsoft-com:office:office"
xmlns:dt="uuid:C2F41010-65B3-11d1-A29F-00AA00C14882"
xmlns="http://www.w3.org/TR/REC-html40">

<head>

<meta name=ProgId content=OneNote.File>
<meta name=Generator content="Microsoft OneNote 15">
</head>

<body lang=en-US style='font-family:Calibri;font-size:11.0pt'>
<!--StartFragment-->

<div style='direction:ltr'>



Scenario | Before (Roslyn edits handling) | After (Roslyn range handling) | UPDATED After numbers (after review feedback)
-- | -- | -- | --
Insert line into   ~1.6k line file | 1228.4ms | 1231.6ms | 1212.6ms
Delete line from   ~1.6k line file | 1121ms | 1254.6ms | 1125.4ms



</div>

<!--EndFragment-->
</body>

</html>

- A pro for Razor with this change is that it at least cleans up the code significantly 😃 
- Deleted unnecessary tests, modified existing tests, and added a test to account for the change.