@ryanbrandenburg 2021-11-22
> Essentially perf-wise this PR has minimal benefit to Razor and mostly benefits Roslyn

This sticks out to me as it doesn't jive with my current understanding of the Razor perf situation. I had it in my head that Roslyn accounted for a significant portion of time spent in non-trivial Razor files (like >50%), and therefore any perf benefit for Roslyn must also be a significant benefit for Razor, right? What do the numbers look like from the Roslyn side?

If this change doesn't rely on the Roslyn one and simplifies the code without paying a significant/any perf cost it seems like we could take it anyway, but the fact that it doesn't improve the Razor performance makes me warry that it will improve Roslyn.