@ryanbrandenburg 2021-11-22
Talked with @allisonchou offline. She explained that Roslyn's perf wins are because they're only doing calculations for the visible area (which is mostly all VS requests) rather than by avoiding the edits algorithm. Because Razor does not limit the range like that we won't get the perf benefit.