@NTaylorMullen 2021-11-22
> Talked with @allisonchou offline. She explained that Roslyn's perf wins are because they're only doing calculations for the visible area (which is mostly all VS requests) rather than by avoiding the edits algorithm. Because Razor does not limit the range like that we won't get the perf benefit.

Hmmm, in relation to [this thread](https://github.com/dotnet/razor-tooling/pull/5769#discussion_r753876026). If we were to only take range and turn off edits as well would that enable us to restrict the requested C# range in a performant way / would it be a net-perf-win?