@ryanbrandenburg 2021-11-22
> > Talked with @allisonchou offline. She explained that Roslyn's perf wins are because they're only doing calculations for the visible area (which is mostly all VS requests) rather than by avoiding the edits algorithm. Because Razor does not limit the range like that we won't get the perf benefit.
> 
> Hmmm, in relation to [this thread](https://github.com/dotnet/razor-tooling/pull/5769#discussion_r753876026). If we were to only take range and turn off edits as well would that enable us to restrict the requested C# range in a performant way / would it be a net-perf-win?

It's possible, if the platform behavior has been changed to not flood us at the beginning. The perf gains from the Razor side are unlikely to be noticeable (unless our edits algo is hurting us a lot more than I imagine), but the Roslyn gains may overtake that. That all assumes we don't get bad startup experience though.