@allisonchou 2021-11-23
> Hmmm, maybe I'm reading into this wrong but don't we drop the range accidentally prior to re-invoking? Aka we construct this "newParams" and then pass it to reinvoke below but I'm not seeing where we update the newParams to lift the range from semanticTokensParams

Oh my goodness, you're totally right 🤦‍♀️ I didn't even notice. I guess this means the numbers in the original post are inaccurate now, they should be a little better than originally reported (I'll try to get some updated numbers soon).

A+ catch, thanks!

> And we can't move to range handling (instead of edits) currently because the platform doesn't support polling semantic colorization changes without document changes right?

Actually, I think the platform will continue to poll us regardless of whether there's changes. @ryanbrandenburg can correct me if I'm wrong, but I think the concern was primarily around perf and startup UI thread usage since the platform will try to queue range requests multiple times. For Roslyn this isn't an issue since I don't believe they block the UI thread and can compute range requests quickly.

In summary, I _think_ we could hypothetically switch to exclusively range handling in Razor but we'd need to do more perf analysis. Filed https://github.com/dotnet/razor-tooling/issues/5775 to track this work.