@ryanbrandenburg 2021-11-23
> Actually, I think the platform will continue to poll us regardless of whether there's changes. @ryanbrandenburg can correct me if I'm wrong, but I think the concern was primarily around perf and startup UI thread usage since the platform will try to queue range requests multiple times. For Roslyn this isn't an issue since I don't believe they block the UI thread and can compute range requests quickly.

This is how I remember it, but that conversation was long enough ago and detail-oriented enough that I'd believe either way.